# whatsapp-clone/whatsappclone/src/main/java/com/devx/whatsappclone/chat/ChatResponse.java
package com.devx.whatsappclone.chat;

import lombok.*;

import java.time.LocalDateTime;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class ChatResponse {

    private String id;
    private String name;
    private long unreadCount;
    private String lastMessage;
    private LocalDateTime lastMessageTime;
    private boolean isRecipientOnline;
    private String senderId;
    private String recipientId;
}