# whatsapp-clone/whatsappclone/src/main/java/com/devx/whatsappclone/user/UserResponse.java
package com.devx.whatsappclone.user;

import lombok.*;

import java.time.LocalDateTime;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class UserResponse {

    private String id;
    private String firstName;
    private String lastName;
    private String email;
    private LocalDateTime lastSeen;
    private Boolean isOnline;
}